pulled out common CN updating code to share with k8s case
